ensure packages are installed after "yum install"
When "yum install <list-of-packages>" fails to install package(s) in the
list, it will simply print a warning, eg, "No package ceph-radosgw
available". It still exits with a "0" exit code in this case, leading
Docker to continue on with the build process as if nothing is wrong.

This can happen if one of the packages is accidentally unavailable in
our Yum repositories.

If we fail to install a package in this list, we need the build process
to fail immediately so we know the container will not be usable.

Rather than only relying on "yum install"'s exit code, use rpm to
double-check that the packages were indeed installed locally.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>